[FEATURE] Replace old result highlighting with a more powerful feature
The ResultHighlightingFeature does no longer use the _all field for highlighting. Instead it is configurable which fields to use - these are written into a (also configurable) concatenated field.

To make this feature work, it needs to be included into
1) Indexer
2) The toplevel collector plus each collector that should add highlight fields
3) The query (already done by default)

and also activated via TypoScript.
For legacy reasons you can use the constant 'settings.result.showHighlighting' which was used previously.