[BUGFIX] Change Toplevel update configuration to hold an array of types for each table
Sometimes it may be necessary to split a table into different types. Without this change they would override each other and direct all updates to one type.

Side effect: If you change a record, it will be scheduled as update for ALL connected types, even if the record does not belong to them (by pid restriction etc.).
However, it will not be indexed as the DataCollector checks all restrictions before adding/updating records. (So it is basically just ugly, but not breaking anything).