[FEATURE] Add HighlightFeature as more efficient replacement of the old ResultHighlightFeature
The new feature uses a mapping option (copy_to) to collect highlight fields - this means no real content duplication and better indexer performance.

The old ResultHighlightFeature is deprecated and will be removed in V3.
Please note that the config option 'strip_tags' is not available for the new feature. There will be an alternative when V3 is released.