[BUGFIX] Forward prepared TYPO3_* env variables to webserver
This is essential for the checks in
SystemEnvironmentBuilder::initializeEnvironment() where TYPO3_PATH_ROOT
and TYPO3_PATH_APP must be properly related to each other. These are
defined in AbstractTestSystem::setTypo3Context() and involve some
processing.

With this the config/var path is always the same both for the test
itself as well as the website accessed by the temporary webserver.