[TASK] Set language aspect for indexing
This is essential for preview rendering which uses Extbase for
retrieving objects. Here we need to set the language aspect since this
is used by the Typo3QuerySettings to prepare language overlays.

Also reset the Extbase persistence between indexing for a language to
ensure objects are retrieved with the correct language.