[BUGFIX] Fall back to first site
In case no site could be resolved in the link building request (e.g. if
a development host is used not covered by site config), we now fall back
to the first site.

This basically reflects the PseudoSite handling which was dropped with
TYPO3v10 which was used since we did construct TSFE with a fixed page
UID "1".