[BUGFIX] Ensure PageArguments for frontend controller
By default the request "routing" attribute contains a SiteRouteResult
set by the SiteResolver middleware. This is normally replaced by
PageArguments by the PageResolver middleware.

Our UriBuilder middleware runs before that and we cannot change the order
here since we need need to be run before the StaticRouteResolver which
runs after the PageResolver middleware.