[TASK] Continue indexing on missing site language
While technically correct we should not fail hard in this case. The
fallback to the site default language basically restores the old
behavior in this case.

A proper fix would be excluding indexers from index languages or
configuring the languages supported by an indexer. But this is beyond
the scope of the current topic.